BudgetApp/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.11.3
pydantic-settings==2.9.1
psycopg2-binary==2.9.10
sqlalchemy>=2.0.36
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
google-generativeai==0.5.4
python-dotenv==1.0.1
httpx==0.27.0